+++ lib/animate.dart
@@ -100,9 +100,8 @@ class Animate extends StatefulWidget with AnimateManager<Animate> {
     Key? key,
     this.child = const SizedBox.shrink(),
     List<Effect>? effects,
-    this.onInit,
-    this.onPlay,
     this.onComplete,
+    this.onPlay,
     this.delay = Duration.zero,
     this.controller,
     this.adapter,
@@ -114,29 +113,14 @@ class Animate extends StatefulWidget with AnimateManager<Animate> {
   /// The widget to apply effects to.
   final Widget child;
 
-  /// Called when the [AnimationController] for this instance is initialized,
-  /// but before it starts playing.
-  /// 
-  /// This typically occurs once when the widget [State] is created, but can
-  /// also happen if the widget is updated with a new [controller].
-  /// 
-  /// Can be used to (for example) loop or reverse
-  /// the animation, but cannot be used to adjust its starting postion or 
-  /// pause it because `AnimationController.forward(from: 0)` is called after
-  /// `onInit` – use [onPlay] instead.
-  /// 
-  /// This would loop the animation, reversing it on each loop:
-  /// ```
-  /// foo.animate(
-  ///   onInit: (controller) => controller.repeat(reverse: true)
-  /// ).fadeIn()
-  /// ```
-  final AnimateCallback? onInit;
+  /// Called when all effects complete. Provides an opportunity to
+  /// manipulate the [AnimationController] (ex. to loop, reverse, etc).
+  final AnimateCallback? onComplete;
 
-  /// Called when the animation begins playing (ie. immediately after
-  /// [AnimationController.forward] is called after [delay]).
+  /// Called when the animation begins playing (ie. after [Animate.delay],
+  /// immediately after [AnimationController.forward] is called).
   /// Provides an opportunity to manipulate the [AnimationController]
-  /// (ex. to loop, reverse, stop, change starting position, etc).
+  /// (ex. to loop, reverse, stop, etc).
   /// 
   /// For example, this would pause the animation at its start:
   /// ```
@@ -144,16 +128,17 @@ class Animate extends StatefulWidget with AnimateManager<Animate> {
   ///   onPlay: (controller) => controller.stop()
   /// ).fadeIn()
   /// ```
+  /// This would loop the animation, reversing it on each loop:
+  /// ```
+  /// foo.animate(
+  ///   onPlay: (controller) => controller.repeat(reverse: true)
+  /// ).fadeIn()
+  /// ```
   final AnimateCallback? onPlay;
-
-  /// Called when all effects complete (ie. the [AnimationController] status
-  /// is [AnimationStatus.completed]). A looping animation will never call 
-  /// `onComplete`.
-  final AnimateCallback? onComplete;
 
   /// A duration to delay before the animation is started. Unlike [Effect.delay],
   /// this is not a part of the overall animation, and only runs once if the
-  /// animation is looped. [onInit] is called after this delay.
+  /// animation is looped. [onPlay] is called after this delay.
   final Duration delay;
 
   /// An external [AnimationController] can optionally be specified. By default
@@ -166,7 +151,7 @@ class Animate extends StatefulWidget with AnimateManager<Animate> {
   /// [ValueNotifierAdapter]).
   ///
   /// If an adapter is provided, then [delay] is ignored, and you should not
-  /// make changes to the [AnimationController] directly (ex. via [onInit])
+  /// make changes to the [AnimationController] directly (ex. via [onPlay])
   /// because it can cause unexpected results.
   final Adapter? adapter;
 
@@ -251,7 +236,6 @@ class _AnimateState extends State<Animate> with SingleTickerProviderStateMixin {
     controller.addStatusListener(_handleAnimationStatus);
     _controller = controller;
     _initAdapter();
-    widget.onInit?.call(_controller);
   }
 
   void _initAdapter() {
@@ -302,7 +286,7 @@ extension AnimateWidgetExtensions on Widget {
     Key? key,
     List<Effect>? effects,
     AnimateCallback? onComplete,
-    AnimateCallback? onInit,
+    AnimateCallback? onPlay,
     Duration delay = Duration.zero,
     AnimationController? controller,
     Adapter? adapter,
@@ -311,7 +295,7 @@ extension AnimateWidgetExtensions on Widget {
         key: key,
         effects: effects,
         onComplete: onComplete,
-        onInit: onInit,
+        onPlay: onPlay,
         delay: delay,
         controller: controller,
         adapter: adapter,