+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - TBD
+### Added
+- `RotateEffect`
+
+### Changed
+- `SwapEffect.builder` now receives a reference to the original child
+- `Animate.onInit` renamed to `onPlay`
+
 ## [0.3.0] - 2022-06-20
 ### Added
 - `.unblur()` extension method